// src/enums.ts
export enum ApiResponseMessage {
  NetworkError = "Network Error",
  BadRequest = "Bad Request",
  Unauthorized = "Unauthorized",
  Forbidden = "Forbidden",
  NotFound = "Resource Not Found",
  InternalServerError = "Internal Server Error",
  UnProcessableEntity = "Unprocessable Entity",
  UnknownError = "An Unknown Error Occurred",
}